updated bmdb

+++ src/main/java/com/bmdb/web/MovieController.java
@@ -3,7 +3,10 @@ package com.bmdb.web;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.bmdb.business.Movie;
 import com.bmdb.db.MovieRepo;
@@ -12,7 +15,7 @@ import com.bmdb.db.MovieRepo;
 @RestController
 @RequestMapping("/api/movies")
 public class MovieController {
-	
+
 	@Autowired
 	private MovieRepo movieRepo;
 
@@ -20,30 +23,52 @@ public class MovieController {
 	public Iterable<Movie> getAll() {
 		return movieRepo.findAll();
 	}
-	
+
 	@GetMapping("/{id}")
 	public Optional<Movie> get(@PathVariable Integer id) {
 		return movieRepo.findById(id);
 	}
-	
+
 	@PostMapping("/")
 	public Movie add(@RequestBody Movie movie) {
 		return movieRepo.save(movie);
 	}
-	
+
 	@PutMapping("/")
 	public Movie update(@RequestBody Movie movie) {
 		return movieRepo.save(movie);
 	}
 
 	@DeleteMapping("/{id}")
-	public void delete(@PathVariable int id) {
-		movieRepo.deleteById(id);
+	public Optional<Movie> delete(@PathVariable int id) {
+		Optional<Movie> movie = movieRepo.findById(id);
+		if (movie.isPresent()) {
+			try {
+				movieRepo.deleteById(id);
+			} catch (DataIntegrityViolationException dive) {
+				// catch dive when movie exists as fk on another table
+				System.err.println(dive.getRootCause().getMessage());
+				throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,
+						"Foreign Key Constraint Issue - movie id: " + id + " is " + "referred to elsewhere.");
+			} catch (Exception e) {
+				e.printStackTrace();
+				throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,
+						"Exception caught during movie delete.");
+			}
+		} else {
+			System.err.println("Movie delete error - no movie found for id:" + id);
+		}
+		return movie;
 	}
-	@GetMapping("/rating/{id}")
+
+	@GetMapping("/rating/{rating}")
 	public Iterable<Movie> getAllByRating(@PathVariable String rating) {
-//		Optional<Movie> movie = movieRepo.findById(id);
-		return movieRepo.findAllByMovieId(rating);
+		return movieRepo.findAllByRating(rating);
+	}
+
+	@GetMapping("/year/{year}")
+	public Iterable<Movie> getAllSinceYear(@PathVariable int year) {
+		return movieRepo.findAllByYearGreaterThanEqual(year);
 	}
 
 }
